requirements.txt
Django==1.8.3
Markdown==2.6.2
PyJWT==1.4.0
PyYAML==3.11
django-braces==1.8.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-gravatar2==1.3.0
django-modeltranslation==0.10
django-rest-swagger==0.3.3
django-sitetree==1.4.0
django-suit==0.2.14
djangorestframework==3.1.3
oauthlib==1.0.0
psycopg2==2.6.1
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.15
#wsgiref==0.1.2
